Move BlockMotion back into Motion
This helps with Code reuse
